requirements.txt
streamlit==1.40.1
pydantic==2.7.4
python-dotenv==1.0.1
pandas==2.2.2
litellm==1.42.6
langgraph==0.2.45
langchain-core==0.3.20
orjson==3.10.7
rich==13.9.2
nicegui
openai>=1.0.0
chromadb>=0.4.0
